third_party/blink/web_tests/external/wpt/trusted-types/navigate-to-javascript-url-001-expected.txt

This is a testharness.js-based test.
[FAIL] Setting window.location to a javascript: URL without a default policy should report a CSP violation instead of executing the JavaScript code.
  assert_equals: expected 4 but got 0
Harness: the test ran to completion.
